Moved graphql to each paragraph
