+++ nextjs/components/paragraphs/ParagraphSidebyside.tsx
@@ -1,9 +1,40 @@
 import React from 'react';
-import { FragmentOf, readFragment } from 'gql.tada';
-import { TextFragment, LinkFragment } from "@/graphql/fragments/misc";
-import { ParagraphSidebysideFragment } from "@/graphql/fragments/paragraph";
+import { FragmentOf, readFragment, graphql } from 'gql.tada';
+import { DateTimeFragment, LanguageFragment, TextFragment, LinkFragment } from "@/graphql/fragments/misc";
+import { MediaUnionFragment } from "@/graphql/fragments/media";
 import { getImage } from '@/components/helpers/Utilities';
 import Sidebyside from '@/components/sidebyside/Sidebyside';
+
+export const ParagraphSidebysideFragment = graphql(`fragment ParagraphSidebysideFragment on ParagraphSidebyside {
+  id
+  created {
+    ...DateTimeFragment
+  }
+  eyebrow
+  langcode {
+    ...LanguageFragment
+  }
+  link {
+    ...LinkFragment
+  }
+  media {
+    ...MediaUnionFragment
+  }
+  sidebysideLayout
+  status
+  sidebysideSummary: summary {
+    ...TextFragment
+  }
+  sidebysideTitle: title
+}`,
+  [
+    DateTimeFragment,
+    LanguageFragment,
+    LinkFragment,
+    MediaUnionFragment,
+    TextFragment,
+  ]
+)
 
 interface ParagraphSidebysideProps {
   paragraph: FragmentOf<typeof ParagraphSidebysideFragment>,
