+++ nextjs/components/paragraphs/ParagraphView.tsx
@@ -1,7 +1,57 @@
-import { FragmentOf, readFragment } from 'gql.tada';
-import { ParagraphViewFragment } from '@/graphql/fragments/paragraph';
+import { FragmentOf, readFragment, graphql } from 'gql.tada';
+import { TextFragment, DateTimeFragment, LanguageFragment, LinkFragment } from "@/graphql/fragments/misc";
+import { MediaUnionFragment } from "@/graphql/fragments/media";
 import RecentCards from '@/components/views/ViewRecentCards';
 import { Card, CardContent } from "@/components/ui/card";
+
+export const ParagraphViewFragment = graphql(`fragment ParagraphViewFragment on ParagraphView {
+  id
+  created {
+    ...DateTimeFragment
+  }
+  langcode {
+    ...LanguageFragment
+  }
+  link {
+    ...LinkFragment
+  }
+  status
+  title
+  viewsRef {
+    __typename
+    ... on RecentCardsArticleCardsResult {
+      id
+      view
+      display
+      results {
+        ... on NodeArticle {
+          id
+          path
+          title
+          media {
+            ...MediaImageFragment
+          }
+          created {
+            ...DateTimeFragment
+          }
+          langcode {
+            ...LanguageFragment
+          }
+          status
+          summary
+        }
+      }
+    }
+  }
+}`,
+  [
+    MediaUnionFragment,
+    DateTimeFragment,
+    LanguageFragment,
+    LinkFragment,
+    TextFragment,
+  ]
+)
 
 interface ParagraphViewProps {
   paragraph: FragmentOf<typeof ParagraphViewFragment>,
