+++ nextjs/graphql/fragments/paragraph.ts
@@ -1,423 +1,16 @@
 import { graphql } from "@/graphql/gql.tada";
-import { TextSummaryFragment, TextFragment, DateTimeFragment, LanguageFragment, LinkFragment } from "./misc";
-import { MediaUnionFragment, SvgMediaFragment } from "./media";
-
-export const ParagraphHeroFragment = graphql(`fragment ParagraphHeroFragment on ParagraphHero {
-  id
-  created {
-    ...DateTimeFragment
-  }
-  heading {
-    ...TextFragment
-  }
-  heroLayout
-  langcode {
-    ...LanguageFragment
-  }
-  link {
-    ...LinkFragment
-  }
-  requiredMedia: media {
-    ...MediaUnionFragment
-  }
-  status
-  summary {
-    ...TextFragment
-  }
-}`,
-  [
-    DateTimeFragment,
-    TextFragment,
-    LanguageFragment,
-    LinkFragment,
-    MediaUnionFragment,
-  ]
-)
-
-export const ParagraphMediaFragment = graphql(`fragment ParagraphMediaFragment on ParagraphMedia {
-  id
-  created {
-    ...DateTimeFragment
-  }
-  langcode {
-    ...LanguageFragment
-  }
-  media {
-    ...MediaUnionFragment
-  }
-  status
-  title
-}`,
-  [
-    DateTimeFragment,
-    LanguageFragment,
-    MediaUnionFragment,
-  ]
-)
-
-export const ParagraphTextFragment = graphql(`fragment ParagraphTextFragment on ParagraphText {
-  id
-  body {
-    ...TextSummaryFragment
-  }
-  textLayout
-  eyebrow
-  created {
-    ...DateTimeFragment
-  }
-  langcode {
-    ...LanguageFragment
-  }
-  link {
-    ...LinkFragment
-  }
-  link2 {
-    ...LinkFragment
-  }
-  status
-  title
-}`,
-  [
-    TextSummaryFragment,
-    DateTimeFragment,
-    LanguageFragment,
-    LinkFragment,
-  ]
-)
-
-export const ParagraphQuoteFragment = graphql(`fragment ParagraphQuoteFragment on ParagraphQuote {
-  id
-  author
-  created {
-    ...DateTimeFragment
-  }
-  jobTitle
-  langcode {
-    ...LanguageFragment
-  }
-  logo {
-    ...SvgMediaFragment
-  }
-  quote
-  status
-  thumb {
-    ...MediaUnionFragment
-  }
-}`,
-  [
-    DateTimeFragment,
-    LanguageFragment,
-    SvgMediaFragment,
-    MediaUnionFragment,
-  ]
-)
-
-export const ParagraphAccordionItemFragment = graphql(`fragment ParagraphAccordionItemFragment on ParagraphAccordionItem {
-  id
-  body {
-    ...TextSummaryFragment
-  }
-  created {
-    ...DateTimeFragment
-  }
-  langcode {
-    ...LanguageFragment
-  }
-  link {
-    ...LinkFragment
-  }
-  status
-  title
-}`,
-  [
-    TextSummaryFragment,
-    DateTimeFragment,
-    LanguageFragment,
-    LinkFragment,
-  ]
-)
-
-export const ParagraphAccordionFragment = graphql(`fragment ParagraphAccordionFragment on ParagraphAccordion {
-  id
-  accordionItem {
-    ...ParagraphAccordionItemFragment
-  }
-  created {
-    ...DateTimeFragment
-  }
-  langcode {
-    ...LanguageFragment
-  }
-  status
-  title
-}`,
-  [
-    ParagraphAccordionItemFragment,
-    DateTimeFragment,
-    LanguageFragment,
-  ]
-)
-
-export const ParagraphCardFragment = graphql(`fragment ParagraphCardFragment on ParagraphCard {
-  id
-  created {
-    ...DateTimeFragment
-  }
-  langcode {
-    ...LanguageFragment
-  }
-  link {
-    ...LinkFragment
-  }
-  media {
-    ...MediaUnionFragment
-  }
-  status
-  summary
-  title
-}`,
-  [
-    DateTimeFragment,
-    LanguageFragment,
-    LinkFragment,
-    MediaUnionFragment,
-  ]
-)
-
-export const ParagraphStatsItemFragment = graphql(`fragment ParagraphStatsItemFragment on ParagraphStatsItem {
-  id
-  created {
-    ...DateTimeFragment
-  }
-  customIcon {
-    ...SvgMediaFragment
-  }
-  langcode {
-    ...LanguageFragment
-  }
-  status
-  statSummary: summary
-  title
-}`,
-  [
-    DateTimeFragment,
-    SvgMediaFragment,
-    LanguageFragment,
-  ]
-)
-
-export const ParagraphCardGroupFragment = graphql(`fragment ParagraphCardGroupFragment on ParagraphCardGroup {
-  id
-  card {
-    __typename
-    ...ParagraphCardFragment
-    ...ParagraphStatsItemFragment
-  }
-  created {
-    ...DateTimeFragment
-  }
-  langcode {
-    ...LanguageFragment
-  }
-  status
-  title
-}`,
-  [
-    ParagraphCardFragment,
-    ParagraphStatsItemFragment,
-    DateTimeFragment,
-    LanguageFragment,
-  ]
-)
-
-export const ParagraphGalleryFragment = graphql(`fragment ParagraphGalleryFragment on ParagraphGallery {
-  id
-  gallerySummary: body {
-    ...TextSummaryFragment
-  }
-  created {
-    ...DateTimeFragment
-  }
-  langcode {
-    ...LanguageFragment
-  }
-  mediaItem {
-    ...MediaUnionFragment
-  }
-  status
-  title
-}`,
-  [
-    TextSummaryFragment,
-    DateTimeFragment,
-    LanguageFragment,
-    MediaUnionFragment,
-  ]
-)
-
-export const ParagraphSidebysideFragment = graphql(`fragment ParagraphSidebysideFragment on ParagraphSidebyside {
-  id
-  created {
-    ...DateTimeFragment
-  }
-  eyebrow
-  langcode {
-    ...LanguageFragment
-  }
-  link {
-    ...LinkFragment
-  }
-  media {
-    ...MediaUnionFragment
-  }
-  sidebysideLayout
-  status
-  sidebysideSummary: summary {
-    ...TextFragment
-  }
-  sidebysideTitle: title
-}`,
-  [
-    DateTimeFragment,
-    LanguageFragment,
-    LinkFragment,
-    MediaUnionFragment,
-    TextFragment,
-  ]
-)
-
-export const ParagraphCarouselItemFragment = graphql(`fragment ParagraphCarouselItemFragment on ParagraphCarouselItem {
-  id
-  created {
-    ...DateTimeFragment
-  }
-  langcode {
-    ...LanguageFragment
-  }
-  media {
-    ...MediaUnionFragment
-  }
-  status
-  summary
-  title
-}`,
-  [
-    DateTimeFragment,
-    LanguageFragment,
-    MediaUnionFragment,
-  ]
-)
-
-export const ParagraphCarouselFragment = graphql(`fragment ParagraphCarouselFragment on ParagraphCarousel {
-  id
-  carouselItem {
-    ...ParagraphCarouselItemFragment
-  }
-  created {
-    ...DateTimeFragment
-  }
-  langcode {
-    ...LanguageFragment
-  }
-  status
-}`,
-  [
-    ParagraphCarouselItemFragment,
-    DateTimeFragment,
-    LanguageFragment,
-  ]
-)
-
-export const ParagraphEmbedFragment = graphql(`fragment ParagraphEmbedFragment on ParagraphEmbed {
-  id
-  created {
-    ...DateTimeFragment
-  }
-  langcode {
-    ...LanguageFragment
-  }
-  script {
-    ...TextFragment
-  }
-  status
-  title
-}`,
-  [
-    DateTimeFragment,
-    LanguageFragment,
-    TextFragment,
-  ]
-);
-
-export const ParagraphNewsletterFragment = graphql(`fragment ParagraphNewsletterFragment on ParagraphNewsletter {
-  id
-  created {
-    ...DateTimeFragment
-  }
-  langcode {
-    ...LanguageFragment
-  }
-  status
-  newsletterTitle: title
-  summary {
-    ...TextFragment
-  }
-}`,
-  [
-    DateTimeFragment,
-    LanguageFragment,
-    TextFragment,
-  ]
-)
-
-export const ParagraphViewFragment = graphql(`fragment ParagraphViewFragment on ParagraphView {
-  id
-  created {
-    ...DateTimeFragment
-  }
-  langcode {
-    ...LanguageFragment
-  }
-  link {
-    ...LinkFragment
-  }
-  status
-  title
-  viewsRef {
-    __typename
-    ... on RecentCardsArticleCardsResult {
-      id
-      view
-      display
-      results {
-        ... on NodeArticle {
-          id
-          path
-          title
-          media {
-            ...MediaImageFragment
-          }
-          created {
-            ...DateTimeFragment
-          }
-          langcode {
-            ...LanguageFragment
-          }
-          status
-          summary
-        }
-      }
-    }
-  }
-}`,
-  [
-    MediaUnionFragment,
-    DateTimeFragment,
-    LanguageFragment,
-    LinkFragment,
-    TextFragment,
-  ]
-)
+import { ParagraphAccordionFragment } from "@/components/paragraphs/ParagraphAccordion";
+import { ParagraphCardGroupFragment } from "@/components/paragraphs/ParagraphCardGroup";
+import { ParagraphHeroFragment } from "@/components/paragraphs/ParagraphHero";
+import { ParagraphMediaFragment } from "@/components/paragraphs/ParagraphMedia";
+import { ParagraphTextFragment } from "@/components/paragraphs/ParagraphText";
+import { ParagraphQuoteFragment } from "@/components/paragraphs/ParagraphQuote";
+import { ParagraphGalleryFragment } from "@/components/paragraphs/ParagraphGallery";
+import { ParagraphSidebysideFragment } from "@/components/paragraphs/ParagraphSidebyside";
+import { ParagraphCarouselFragment } from "@/components/paragraphs/ParagraphCarousel";
+import { ParagraphEmbedFragment } from "@/components/paragraphs/ParagraphEmbed";
+import { ParagraphNewsletterFragment } from "@/components/paragraphs/ParagraphNewsletter";
+import { ParagraphViewFragment } from "@/components/paragraphs/ParagraphView";
 
 export const ParagraphUnionFragment = graphql(`
   fragment ParagraphUnionFragment on ParagraphUnion {
